Add SMB Volume Driver
The purpose of this blueprint is to add an SMB Cinder volume driver.
It will include the minimum set of features required by the Juno release.

This driver will support Linux distributions (using Samba).
Another driver based on it will be added for Windows Server 2012
(R2 included) as well.

Volumes will be exported through SMB shares as disk images,
thus having a similar flow with the NFS driver.

Change-Id: Id6ca4a8f7caea4da2aa406d02b57db37135b8316