Windows SMB Volume Driver
The purpose of this blueprint is to add an SMB Cinder volume driver.
It will include the minimum set of features required by the Juno
release.

This driver will support Windows Server 2012 (R2 included) and it
will be based on the volume driver that will be available for Linux
distributions, which is drafted here:

https://blueprints.launchpad.net/cinder/+spec/smbfs-volume-driver

Volumes will be exported through SMB shares as disk images,
thus having a similar flow with the NFS driver.

Change-Id: I594fb303d1cdba76794ec10917ed21ea3371cf49