Get supported volume type extra specs keys
The main purpose of this spec is to help improve Horizon's user
experience with the volume type extra specs management processes. This
spec proposes that Cinder provides a list of possible volume type extra
specs keys for users to choose from when creating or editing extra
specs. The current process in Horizon provides no useful help in terms
of what the possible extra specs keys and values might be. Users have to
know the exact spellings ahead of time. Having a list of possible keys
for users to choose from will help lessen the time users spend on this
task.

DocImpact
blueprint: get-volume-type-extra-specs
Change-Id: I0aa874bff02a5e86eb8237260aafaeffd7408afd