Implement force_detach for safe cleanup
For volumes stuck in 'attaching' or 'detaching' , there is no safe way
to cleanup that involves the backend storage. Using python-cinderclient
'reset-state' will only change the Cinder database, and may leave the
volume exported to the compute host, and may leave an entry in Nova's
database that prevents the volume from being re-used. This
blueprint+spec outlines the changes to the cinder side to allow a
force_detach.

Change-Id: I876741e392c3477af2f3680ca67b0576711a8cfa