Support clone_image from Glance cinder backend
When Glance cinder store is enabled, the glance image can store
locations of cinder volume. When a raw image is cinder volume
backed, we can efficiently create a new volume from the image by
cloning the backing volume, instead of downloading the image from
Glance. We can also make upload-to-image from a volume efficient
by creating a cloned volume and add its location to an image.

Change-Id: Ia930672ab941dd0719a407c760518649a66b3864
blueprint: clone-image-in-glance-cinder-backend