Replication with More Granularity (Tiramisu)
In Mitaka, replication v2.1 (Cheesecake) was implemented to support a
disaster recovery scenario when an entire host can be failed over to
a secondary site. Cheesecake API was admin facing.

At the Mitaka mid-cycle meetup, Cinder team decided that after Cheesecake
is implemented, Tiramisu will be the next step. Tiramisu is the code name
for a new set of replication APIs with more granularity. Tiramisu API will
be tenant facing. It gives tenant more control on what should be replicated
together, i.e., a group of volumes.

Note: This spec does not address replication at the single volume level
because it is built upon generic volume groups.

Change-Id: I377fd84bd19ae78dc2e952a7846573546bc83f2c
blueprint: replication-cg