List volumes/snapshots for manage-existing
Cinder currently has the ability to take over the management of existing
volumes and snapshots ("manage existing") and to relinquish management of
volumes and snapshots ("unmanage"). The API to manage an existing volume takes
a reference, which is a driver-specific string that is used to identify the
volume on the storage backend. This process is currently confusing and error
prone. This spec's purpose is to detail APIs for listing volumes and snapshots
available for management to make this flow more user-friendly. Allowing the
listing of available volumes/snapshots will allow for the creation of an
easy-to-use GUI to migrate existing volumes to a new OpenStack environment, or
to recover in the case the Cinder database is lost.

blueprint: list-manage-existing
Change-Id: Ie2a22deaac61ffe3ca69f62012d4cf894638c83c