Update Job Distribution for A/A Specs
During Austin Summit it was decided that we should change approved specs
approach to Job Distribution for Active-Active configurations.

This patch updates the specs to use the concept of cluster instead of
using the host and adding the concept of nodes as agreed during the
design session.

It also moves the spec to Ocata to make sure there's no confusion on
whether this is implemented or not.

Change-Id: I5f8194a52ff249d85fdf4ac9cfb184a9f33e1ccc
Implements: blueprint cinder-volume-active-active-support