Update cheesecake promotion specification
This version addresses several concerns raised
about the original spec to allow for drivers to
have a chance to update the backend when
replication state and/or cinder.conf changes.

It also allows for modifying more of the db
entries related to replication for the services
when doing promotion.

Blueprint replication-backend-promotion

Change-Id: I268f138fc79ec43d5d1842c14a11a78d9062d49e