Fix spec format errors
Apparently in the window between switching from Jenkins to Zuul, some
specs that were merged by zuul contained some format errors that
Jenkins tests for but zuul does not. So now that we've switched back
to Jenkins, and new specs trying to merge are getting test failures
due to these other specs.

This addresses the failure seen as well as a few other doc8 issues
pointed out in case they cause issues as well.

Change-Id: I92554c5481417e5d4e0cc931a1f6efdbee9b8733