Clean up required packages
This patch does two things to clean up gate package requirements,
adding a bindep file and moving packages to the right location.

Bindep controls which system packages are installed. If a project
does not have a bindep.txt file present, the gate jobs will fall
back to a common list of packages. These include things like
mariadb and various clients that are not needed for specs repos.
Added a basic bindep.txt file with just the things we need to
ensure are present for building the docs in specs.

There are also plans to stop looking at requirements.txt for specs
since these repos do not install Python code. The two locations
that are looked at are test-requirements.txt for running included
tests, and doc/requirements.txt for anything needed for doc builds
such as sphinx and plugins.

Change-Id: I3a222d7eec32b0d5ae6ef4ad20b34603931e13d6