Really open cinder-specs for Rocky
The previous commit to open cinder-specs for Rocky
doesn't appear to have worked:
2353d9a0d5dd24ae9d0b61c847fd21aad51ea012

It appears that just putting .gitignore in the directory
doesn't cause git to actually create the directory. I can see
that the patch merged but the directory doesn't appear in
subsequent pulls of cinder-specs.  Going back to the
method of using .placeholder here.

Change-Id: Ib94315f73f59aa9c3910731d6c166d6b020bf65c