Update Provisioning Improvements
During the implementation phase of the Provisioning Improvements specs
an error in the calculation formula of the merged spec was discovered,
so the patch used the formula from a previous revision of the spec
patch.

The implementation also diverged in not adding a new configuration
option but overloading the existing one to support the `auto` option.

This patch updates the Provisioning Improvements specs to match the
implementation from change id If30bb6276f58532c0f78ac268544a8804008770e

Change-Id: Id4b7aa252ad0404d3038a1279e0d673a0809c7e8