Add untargeted folder for specs
It was discussed at the Cinder Mid-Cycle that we
needed something to do with approved specs that were
still relevant but not being worked.  The recommendation
was to add an untargeted folder like the one that Glance
uses.  This patch does this.

I also discovered that the last two sections in
the index page were not displaying due to
incorrect RST format.  This patch also fixes
that.

Change-Id: I3f909742501ab3a3f36ba9eee4ede0b814f76f2d