Move dynamic reconfiguration spec to untargeted
Now that we have the new untargeted folder for specs I
have a spec to make use of it.  We have approved
implementing dynamic reconfiguration (being able to change
configuration options from cinder.conf without restarting).

This is still a function that we would like to plan to get in
at some point but don't have anyone to pick up the work.  So,
I am moving it to untargeted and hope someone might pick it up.

Change-Id: I7cae8e5397659ce8ff4ab3ecb205a8228b9e7410