Move generic backup spec to untargeted
We have been sliding the generic backup spec along for
a while and I believe that Ivan wasn't planning to get this
in for train.  As a result I am going to move this to
untargeted and we can put it into a release when we know
the code is closer to landing.

Change-Id: Ic13e52cdb6d8dd7c83ead34fc2f80b7b03065f5e