New Quota System
Cinder quotas have been a constant pain for operators and cloud users.
This spec proposes a new quota system to resolve inconsistencies in
tracking quota usage.

This spec proposes reduce the usage of the reservations to a minimum and
have 2 quota drivers, one that leverages the DB engine to dynamically
count resources and another that stores the in-use and reserved counters
in the DB similarly to what we do today.

Change-Id: I3773be238a9edfc971d2b81135b23abfd3c974d7
Implements: blueprint count-resource-to-check-quota